+++ Cleaner/ProductCleaner.php
@@ -7,6 +7,8 @@ use Pim\Bundle\MagentoConnectorBundle\Validator\Constraints\HasValidCredentials;
 use Pim\Bundle\CatalogBundle\Manager\ChannelManager;
 use Pim\Bundle\MagentoConnectorBundle\Guesser\WebserviceGuesser;
 use Pim\Bundle\CatalogBundle\Manager\ProductManager;
+use Pim\Bundle\MagentoConnectorBundle\Webservice\SoapCallException;
+use Oro\Bundle\BatchBundle\Item\InvalidItemException;
 
 /**
  * Magento product cleaner
@@ -19,19 +21,39 @@ use Pim\Bundle\CatalogBundle\Manager\ProductManager;
  */
 class ProductCleaner extends Cleaner
 {
-    const DO_NOTHING = 'do_nothing';
-    const DISABLE    = 'disable';
-    const DELETE     = 'delete';
+    /**
+     * @Assert\NotBlank(groups={"Execution"})
+     */
+    protected $channel;
+
+    /**
+     * get channel
+     *
+     * @return string channel
+     */
+    public function getChannel()
+    {
+        return $this->channel;
+    }
+
+    /**
+     * Set channel
+     *
+     * @param string $channel channel
+     *
+     * @return AbstractProcessor
+     */
+    public function setChannel($channel)
+    {
+        $this->channel = $channel;
+
+        return $this;
+    }
 
     /**
      * @var string
      */
     protected $notCompleteAnymoreAction;
-
-    /**
-     * @var string
-     */
-    protected $notInPimAnymoreAction;
 
     /**
      * get notCompleteAnymoreAction
@@ -58,41 +80,18 @@ class ProductCleaner extends Cleaner
     }
 
     /**
-     * get notInPimAnymoreAction
-     *
-     * @return string notInPimAnymoreAction
-     */
-    public function getNotInPimAnymoreAction()
-    {
-        return $this->notInPimAnymoreAction;
-    }
-
-    /**
-     * Set notInPimAnymoreAction
-     *
-     * @param string $notInPimAnymoreAction notInPimAnymoreAction
-     *
-     * @return ProductCleaner
-     */
-    public function setNotInPimAnymoreAction($notInPimAnymoreAction)
-    {
-        $this->notInPimAnymoreAction = $notInPimAnymoreAction;
-
-        return $this;
-    }
-
-    /**
-     * @param ChannelManager    $channelManager
      * @param WebserviceGuesser $webserviceGuesser
+     * @param ChannelManager    $channelManager
      * @param ProductManager    $productManager
      */
     public function __construct(
-        ChannelManager $channelManager,
         WebserviceGuesser $webserviceGuesser,
+        ChannelManager $channelManager,
         ProductManager $productManager
     ) {
-        parent::__construct($channelManager, $webserviceGuesser);
+        parent::__construct($webserviceGuesser);
 
+        $this->channelManager = $channelManager;
         $this->productManager = $productManager;
     }
 
@@ -108,10 +107,14 @@ class ProductCleaner extends Cleaner
         $pimProducts      = $this->getPimProductsSkus();
 
         foreach ($magentoProducts as $product) {
-            if (!in_array($product['sku'], $pimProducts)) {
-                $this->handleProductNotInPimAnymore($product);
-            } elseif (!in_array($product['sku'], $exportedProducts)) {
-                $this->handleProductNotCompleteAnymore($product);
+            try {
+                if (!in_array($product['sku'], $pimProducts)) {
+                    $this->handleProductNotInPimAnymore($product);
+                } elseif (!in_array($product['sku'], $exportedProducts)) {
+                    $this->handleProductNotCompleteAnymore($product);
+                }
+            } catch (SoapCallException $e) {
+                throw new InvalidItemException($e->getMessage(), array(json_encode($product)));
             }
         }
     }
@@ -200,21 +203,17 @@ class ProductCleaner extends Cleaner
                     'type'    => 'choice',
                     'options' => array(
                         'choices'  => array(
-                            self::DO_NOTHING => self::DO_NOTHING,
-                            self::DISABLE => self::DISABLE,
-                            self::DELETE => self::DELETE
+                            Cleaner::DO_NOTHING => Cleaner::DO_NOTHING,
+                            Cleaner::DISABLE    => Cleaner::DISABLE,
+                            Cleaner::DELETE     => Cleaner::DELETE
                         ),
                         'required' => true
                     )
                 ),
-                'notInPimAnymoreAction' => array(
+                'channel'      => array(
                     'type'    => 'choice',
                     'options' => array(
-                        'choices'  => array(
-                            self::DO_NOTHING => self::DO_NOTHING,
-                            self::DISABLE => self::DISABLE,
-                            self::DELETE => self::DELETE
-                        ),
+                        'choices'  => $this->channelManager->getChannelChoices(),
                         'required' => true
                     )
                 )