+++ Normalizer/CategoryNormalizer.php
@@ -44,18 +44,18 @@ class CategoryNormalizer extends AbstractNormalizer
 
         //For each storeview, we update the product only with localized attributes
         foreach ($this->getPimLocales($context['channel']) as $locale) {
-            $storeViewCode = $this->getStoreViewCodeForLocale(
+            $storeView = $this->getStoreViewForLocale(
                 $locale->getCode(),
                 $context['magentoStoreViews'],
                 $context['storeViewMapping']
             );
 
             //If a locale for this storeview exist in PIM, we create a translated product in this locale
-            if ($storeViewCode) {
+            if ($storeView) {
                 $normalizedCategory['variation'][] = $this->getNormalizedVariationCategory(
                     $object,
                     $locale->getCode(),
-                    $storeViewCode
+                    $storeView['code']
                 );
             }
         }
@@ -79,7 +79,7 @@ class CategoryNormalizer extends AbstractNormalizer
             'variation' => array()
         );
 
-        if ($this->magentoCategoryExist($category, $context['magentoCategories'], $context['magentoUrl'])) {
+        if ($this->magentoCategoryExists($category, $context['magentoCategories'], $context['magentoUrl'])) {
             $normalizedCategory['update'][] = $this->getNormalizedUpdateCategory(
                 $category,
                 $context
@@ -103,7 +103,7 @@ class CategoryNormalizer extends AbstractNormalizer
      *
      * @return boolean
      */
-    protected function magentoCategoryExist(CategoryInterface $category, array $magentoCategories, $magentoUrl)
+    protected function magentoCategoryExists(CategoryInterface $category, array $magentoCategories, $magentoUrl)
     {
         return ($magentoCategoryId = $this->getMagentoCategoryId($category, $magentoUrl)) !== null &&
             isset($magentoCategories[$magentoCategoryId]);